securedendpoints

+++ server/src/main/java/com/codecool/teamup/service/FeedbackService.java
@@ -2,9 +2,13 @@ package com.codecool.teamup.service;
 
 import com.codecool.teamup.model.feedback.Feedback;
 import com.codecool.teamup.model.feedback.FeedbackDTO;
+import com.codecool.teamup.model.user.MyUser;
 import com.codecool.teamup.model.user.UserEntity;
 import com.codecool.teamup.repository.FeedbackRepository;
 import com.codecool.teamup.repository.UserRepository;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -47,6 +51,17 @@ public class FeedbackService {
     }
 
     public void deleteFeedback(long feedbackId) {
-        feedbackRepository.deleteById(feedbackId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        long userId = ((MyUser) authentication.getPrincipal()).getId();
+        Optional<Feedback> optionalFeedback = feedbackRepository.findById(feedbackId);
+        if (optionalFeedback.isPresent()) {
+            if (optionalFeedback.get().getUser().getId() == userId || authentication.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) {
+                feedbackRepository.deleteById(feedbackId);
+            } else {
+                throw new RuntimeException("Unauthorized feedback request");
+            }
+        } else {
+            throw new RuntimeException("Feedback not found");
+        }
     }
 }
